+++ bot/__init__.py
@@ -1,5 +1,13 @@
 from .error_handling import on_app_command_error
 from .load_env import load_env_bot_token
+from .outbot_custom_setup import custom_setup
 
 # DO NOT ADD "from .outbot import OutBot"
 from .run_outbot import run_bot
+
+__all__: list[str] = [
+    "custom_setup",
+    "load_env_bot_token",
+    "on_app_command_error",
+    "run_bot",
+]