+++ bot/error_handling.py
@@ -31,4 +31,3 @@ async def on_app_command_error(
 
         else:
             await interaction.response.send_message(ERROR_MESSAGE, ephemeral=True)
-        print(error)