+++ bot/main.py
@@ -1,10 +1,8 @@
-from .error_handling import on_app_command_error
 from .load_env import load_env_bot_token
+from .outbot_custom_setup import custom_setup
 from .run_outbot import run_bot
-from utils import custom_setup
 
 if __name__ == "__main__":
     outbot = custom_setup()
-    outbot.tree.on_error = on_app_command_error
     load_env_bot_token()
     run_bot()