+++ bot/outbot.py
@@ -2,6 +2,8 @@ import os
 
 from discord.ext import commands
 
+from .error_handling import on_app_command_error
+
 
 class OutBot(commands.Bot):
     """OutBot's custom setup class."""
@@ -13,3 +15,5 @@ class OutBot(commands.Bot):
         for cog in find_cogs:
             if cog.endswith("cog.py"):
                 await self.load_extension(f"cogs.{cog[:-3]}")
+
+        self.tree.on_error = on_app_command_error