+++ cogs/__init__.py
@@ -1,8 +1,19 @@
 from .developer_cog import DeveloperCommands
-from .fun_cog import FreeNitroButton, FunCommands
+from .fun_cog import FunCommands
 from .general_cog import GeneralCommands
-from .information_cog import BotPingButton, InformationCommands
+from .information_cog import InformationCommands
 from .links_cog import LinksCommands
 from .privacy_cog import PrivacyCommands
 from .rules_cog import RulesCommands
-from .support_cog import ReportButton, SupportCommands
+from .support_cog import SupportCommands
+
+__all__: list[str] = [
+    "DeveloperCommands",
+    "FunCommands",
+    "GeneralCommands",
+    "InformationCommands",
+    "LinksCommands",
+    "PrivacyCommands",
+    "RulesCommands",
+    "SupportCommands",
+]