+++ utils/__init__.py
@@ -9,12 +9,34 @@ from .bot_info import (
     LOG_RETENTION,
     OUTBOT_INVITE_LINK,
     OUTBOT_LICENSE,
-    OUTMYTH_YOUTUBE_CHANNEL_LINK,
     PRIVACY_POLICY,
     SECURITY_POLICY,
     TERMS_OF_SERVICE,
 )
 from .error_message import ERROR_MESSAGE
-from .outbot_custom_setup import custom_setup
 from .profanity import send_censor_word_warning
 from .report_embeds import ReportEmbedMessages
+from .views import BotPingButton, FreeNitroButton, ReportButtons, ReportDropdown
+
+__all__: list[str] = [
+    "BOT_VERSION",
+    "CODE_OF_CONDUCT",
+    "CONTRIBUTING_POLICY",
+    "DATE_CREATED",
+    "DEVELOPER",
+    "DISCORD_SERVER_INVITE_LINK",
+    "ERROR_MESSAGE",
+    "GITHUB_LINK",
+    "LOG_RETENTION",
+    "OUTBOT_INVITE_LINK",
+    "OUTBOT_LICENSE",
+    "PRIVACY_POLICY",
+    "SECURITY_POLICY",
+    "TERMS_OF_SERVICE",
+    "BotPingButton",
+    "FreeNitroButton",
+    "ReportButtons",
+    "ReportDropdown",
+    "ReportEmbedMessages",
+    "send_censor_word_warning",
+]