+++ utils/error_message.py
@@ -1,3 +1,5 @@
+from typing import Final
+
 ERROR_MESSAGE: Final[str] = (
     "# Something went wrong :(\nAn unexpected error occurred, please open a ticket."
 )