+++ documentation/CHANGELOG.md
@@ -1,8 +1,14 @@
 # OutBot Bug Fix - 25 September 2026
 
+## Update 0.5.11
+
+- Fixed rate limit message not being sent.
+
+# OutBot Bug Fix - 25 September 2026
+
 ## Update 0.5.10
 
-- Fixed dsicord developer id not being able to be none.
+- Fixed discord developer id not being able to be none.
 
 # OutBot Bug Fix - 22 September 2026